Step One: Enable Auto-configuration
By default, Spring Boot with auto-configure Spring Security to protect
all endpoints. It will also create a user with a generated password.

So that you could better envision life without the auto-configuration,
this was manually shut off in the project.

In a real project, all that's necessary to get this behavior is to add
the spring-boot-starter-security dependency.