Step One: Allowing CORS at an Endpoint
Adding the @CrossOrigin annotation to a Spring MVC mapping opens up
calling that method to any Origin.

Calling the `cors` method on the Spring Security DSL configures Spring
Security to allow the preflight request without the user first being
authenticated.

For HTTP Basic to work in a Single-Page Application, you need to
configure CORS to the cross-origin to send credentials. This should only
be done in a high-trust situation as this would allow malicious users to
induce the client application to send cookies and bypass CSRF
restrictions.