Step Four: Programmatic Security
Spring Security encourages a declaration security authorization model.

However, at times it may be necessary to make an authorization decision
programmatically, like in this case when some information is withheld
if the authentication doesn't have the appropriate authority.

Note that there are declarative ways to achieve this, too. This commit
demonstrates Spring Security's flexibility.